:recycle:  refactor: 작성하기 페이지일 경우 버튼 분리
